src/backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.6
deepdiff==7.0.1
Deprecated==1.2.14
docstring_parser==0.16
emoji==2.12.1
exceptiongroup==1.2.1
filetype==1.2.0
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-aiplatform==1.52.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
idna==3.7
importlib-metadata==7.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-google-vertexai==1.0.4
langchain-postgres==0.0.6
langchain-text-splitters==0.2.0
langdetect==1.0.9
langsmith==0.1.63
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-asyncio==0.45b0
opentelemetry-instrumentation-aws-lambda==0.45b0
opentelemetry-instrumentation-dbapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-instrumentation-sqlite3==0.45b0
opentelemetry-instrumentation-urllib==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-test-utils==0.45b0
opentelemetry-util-http==0.45b0
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pgvector==0.2.5
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
psycopg-pool==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pypdf==4.2.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-iso639==2024.4.27
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.9.1
regex==2024.5.15
requests==2.32.2
rsa==4.9
shapely==2.0.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
tabulate==0.9.0
tenacity==8.3.0
tomli==2.0.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
unstructured==0.14.2
unstructured-client==0.22.0
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
